requirements.txt
aiogram==3.*
aiohttp
python-dotenv
requests ~= 2.32.3
docker ~= 7.1.0